+++ packages/typespec-rust/test/spector/azure/resource-manager/multi-service-older-versions/tests/multi_service_older_versions_test.rs
@@ -38,11 +38,10 @@ impl TokenCredential for FakeTokenCredential {
     }
 }
 
-fn create_client(api_version: &str) -> CombinedClient {
+fn create_client() -> CombinedClient {
     CombinedClient::new(
         "http://localhost:3000",
         Arc::new(FakeTokenCredential::new("fake_token".to_string())),
-        api_version.to_string(),
         "00000000-0000-0000-0000-000000000000".to_string(),
         None,
     )
@@ -51,7 +50,7 @@ fn create_client(api_version: &str) -> CombinedClient {
 
 #[tokio::test]
 async fn virtual_machine_get() {
-    let client = create_client("2024-11-01");
+    let client = create_client();
     let resp = client
         .get_combined_virtual_machines_client()
         .get("test-rg", "vm-old1", None)
@@ -88,7 +87,7 @@ async fn virtual_machine_create_or_update() {
         ..Default::default()
     };
 
-    let client = create_client("2024-11-01");
+    let client = create_client();
     let poller = client
         .get_combined_virtual_machines_client()
         .create_or_update("test-rg", "vm-old1", resource.try_into().unwrap(), None)
@@ -117,7 +116,7 @@ async fn virtual_machine_create_or_update() {
 
 #[tokio::test]
 async fn disk_get() {
-    let client = create_client("2024-03-02");
+    let client = create_client();
     let resp = client
         .get_combined_disks_client()
         .get("test-rg", "disk-old1", None)
@@ -154,7 +153,7 @@ async fn disk_create_or_update() {
         ..Default::default()
     };
 
-    let client = create_client("2024-03-02");
+    let client = create_client();
     let poller = client
         .get_combined_disks_client()
         .create_or_update("test-rg", "disk-old1", resource.try_into().unwrap(), None)